+++ packages/protocol/scripts/bash/verify-release.sh
@@ -9,6 +9,7 @@ set -euo pipefail
 # -b: Branch containing smart contracts that are proposed to be deployed.
 # -n: The network to check
 # -p: The proposal JSON file
+# -i: Path to the data needed to initialize contracts.
 # -f: Boolean flag to indicate if the Forno service should be used to connect to
 # the network
 # -l: Path to a file to which logs should be appended
@@ -17,12 +18,14 @@ BRANCH=""
 NETWORK=""
 PROPOSAL=""
 FORNO=""
+INITIALIZE_DATA=""
 LOG_FILE="/dev/null"
 
-while getopts 'b:n:p:fl:' flag; do
+while getopts 'b:n:p:fl:i:' flag; do
   case "${flag}" in
     b) BRANCH="${OPTARG}" ;;
     n) NETWORK="${OPTARG}" ;;
+    i) INITIALIZE_DATA="--initialize_data ../../${OPTARG}" ;;
     p) PROPOSAL="${OPTARG}" ;;
     f) FORNO="--forno" ;;
     l) LOG_FILE="${OPTARG}" ;;
@@ -35,7 +38,7 @@ done
 [ -z "$PROPOSAL" ] && echo "Need to set the proposal via the -p flag" && exit 1;
 
 BUILD_DIR=$(echo build/$(echo $BRANCH | sed -e 's/\//_/g'))
-git fetch --all --tags >> $LOG_FILE 
+git fetch --all --tags >> $LOG_FILE
 echo " - Checkout source code at $BRANCH"
 git checkout $BRANCH 2>$LOG_FILE > $LOG_FILE
 rm -rf build/contracts
@@ -52,4 +55,4 @@ git checkout - > $LOG_FILE
 echo " - Build verification script ..."
 yarn build > $LOG_FILE
 echo " - Run verification script ..."
-yarn run truffle exec ./scripts/truffle/verify-bytecode.js --network $NETWORK --build_artifacts $BUILD_DIR/contracts --proposal "../../$PROPOSAL" $FORNO
+yarn run truffle exec ./scripts/truffle/verify-bytecode.js --network $NETWORK --build_artifacts $BUILD_DIR/contracts --proposal "../../$PROPOSAL" $FORNO $INITIALIZE_DATA